+++ src/core/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from . import api
+
+app_name = "core"
+
+urlpatterns = [
+    path("about/", api.about),
+]