+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 ## [Unreleased]
 
+## [2] - 2020-03-03
+### Added
+- django shell_plus
+
 ## [1] - 2020-03-02
 ### Changed
 - Improve manage.sh